helper functions to generate inputs moving to a test package inside circuits to make them available for following circuits
